@@ -10,22 +10,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0-next.0",
-    "@angular/common": "~13.0.0-next.0",
-    "@angular/compiler": "~13.0.0-next.0",
-    "@angular/core": "~13.0.0-next.0",
-    "@angular/forms": "~13.0.0-next.0",
-    "@angular/platform-browser": "~13.0.0-next.0",
-    "@angular/platform-browser-dynamic": "~13.0.0-next.0",
-    "@angular/router": "~13.0.0-next.0",
+    "@angular/animations": "^14.2.1",
+    "@angular/common": "^14.2.1",
+    "@angular/compiler": "^14.2.1",
+    "@angular/core": "^14.2.1",
+    "@angular/forms": "^14.2.1",
+    "@angular/platform-browser": "^14.2.1",
+    "@angular/platform-browser-dynamic": "^14.2.1",
+    "@angular/router": "^14.2.1",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.0",
-    "@angular/cli": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0-next.0",
+    "@angular-devkit/build-angular": "^14.2.2",
+    "@angular/cli": "^14.2.2",
+    "@angular/compiler-cli": "^14.2.1",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.10.0",
@@ -34,6 +34,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.8.3"
   }
 }
